requirements.txt
aiohttp>=3.8.4
aiortc>=1.5.0
av>=10.0.0
numpy>=1.24.0
psutil>=5.9.0
pyyaml>=6.0
v4l2-python3>=0.3.2